# shortener_app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "URL Shortener API",
        "description": "API for shortening URLs and managing shortened URLs",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Read Root",
                "operationId": "read_root__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/{url_key}": {
            "get": {
                "summary": "Forward To Target Url",
                "operationId": "forward_to_target_url__url_key__get",
                "parameters": [
                    {
                        "name": "url_key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Url Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/url": {
            "post": {
                "tags": [
                    "url"
                ],
                "summary": "Create Url",
                "operationId": "create_url_url_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/URLBase"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/URLInfo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/{secret_key}": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Administration Info",
                "operationId": "administration_info_admin__secret_key__get",
                "parameters": [
                    {
                        "name": "secret_key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Secret Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/URLInfo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "Delete Url",
                "operationId": "delete_url_admin__secret_key__delete",
                "parameters": [
                    {
                        "name": "secret_key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Secret Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "URLBase": {
                "properties": {
                    "target_url": {
                        "type": "string",
                        "title": "Target Url"
                    },
                    "custom_key": {
                        "type": "string",
                        "title": "Custom Key"
                    }
                },
                "type": "object",
                "required": [
                    "target_url"
                ],
                "title": "URLBase"
            },
            "URLInfo": {
                "properties": {
                    "target_url": {
                        "type": "string",
                        "title": "Target Url"
                    },
                    "custom_key": {
                        "type": "string",
                        "title": "Custom Key"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "clicks": {
                        "type": "integer",
                        "title": "Clicks"
                    },
                    "url": {
                        "type": "string",
                        "title": "Url"
                    },
                    "admin_url": {
                        "type": "string",
                        "title": "Admin Url"
                    },
                    "qr_code": {
                        "type": "string",
                        "title": "Qr Code",
                        "description": "Base64 encoded QR code image for the URL"
                    }
                },
                "type": "object",
                "required": [
                    "target_url",
                    "is_active",
                    "clicks",
                    "url",
                    "admin_url"
                ],
                "title": "URLInfo"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-KEY"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}